pyproject.toml
[project]
name = "wanderly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-experimental>=0.3.4",
    "langchain-google-community[places]>=2.0.7",
    "langchain-openai>=0.3.32",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.6",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "setuptools>=80.9.0",
    "streamlit>=1.49.1",
    "uvicorn>=0.35.0",
]
